fix: fix icon styling in Searchbar
